agents/requirements.txt
# Nexus QA - CrewAI Agents Dependencies

# CrewAI Framework
crewai>=0.86.0
crewai-tools>=0.17.0

# LLM Providers
openai>=1.59.0
# google-generativeai>=0.8.0  # Opsiyonel: Gemini için

# FastAPI for REST API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0

# Test Frameworks
playwright>=1.49.0
pytest>=8.3.0
pytest-playwright>=0.6.0
requests>=2.32.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.28.0
aiohttp>=3.11.0
asyncio>=3.4.3

# Code Analysis
pylint>=3.3.0
black>=24.10.0

# NLP Enhancement (Sembi IQ Style)
spacy>=3.7.0
# Run: python -m spacy download tr_core_news_sm
transformers>=4.40.0
torch>=2.2.0
nltk>=3.8.1
scikit-learn>=1.4.0

